#42 refactor: add content model
