+++ src/main/java/com/kairos/context_engine/application/use_case/GenerateSourceContextUseCase.java
@@ -2,9 +2,9 @@ package com.kairos.context_engine.application.use_case;
 
 import com.kairos.context_engine.application.command.GenerateSourceContextCommand;
 import com.kairos.context_engine.domain.port.embedding.EmbeddingProvider;
-import com.kairos.context_engine.domain.model.Chunk;
+import com.kairos.context_engine.domain.model.content.Chunk;
 import com.kairos.context_engine.domain.model.KnowledgeTriple;
-import com.kairos.context_engine.domain.model.Source;
+import com.kairos.context_engine.domain.model.content.Source;
 import com.kairos.context_engine.domain.model.Triple;
 import com.kairos.context_engine.domain.port.extraction.TripleExtractor;
 import com.kairos.context_engine.domain.port.repository.ChunkRepository;
