+++ src/main/java/com/kairos/context_engine/application/use_case/SearchSourceUseCase.java
@@ -3,7 +3,7 @@ package com.kairos.context_engine.application.use_case;
 import com.kairos.context_engine.application.query.SearchSourceQuery;
 import com.kairos.context_engine.domain.port.embedding.EmbeddingProvider;
 import com.kairos.context_engine.domain.port.graph.KnowledgeGraphSearch;
-import com.kairos.context_engine.domain.model.Chunk;
+import com.kairos.context_engine.domain.model.content.Chunk;
 import com.kairos.context_engine.domain.model.KnowledgeTriple;
 import com.kairos.context_engine.domain.model.SearchResult;
 import com.kairos.context_engine.domain.port.semantic.SemanticSearch;
