+++ src/main/java/com/kairos/context_engine/domain/port/graph/KnowledgeGraphSearch.java
@@ -1,6 +1,6 @@
 package com.kairos.context_engine.domain.port.graph;
 
-import com.kairos.context_engine.domain.model.Chunk;
+import com.kairos.context_engine.domain.model.content.Chunk;
 import com.kairos.context_engine.domain.model.KnowledgeTriple;
 
 import java.util.List;
