+++ src/main/java/com/kairos/context_engine/domain/port/repository/ChunkRepository.java
@@ -1,6 +1,6 @@
 package com.kairos.context_engine.domain.port.repository;
 
-import com.kairos.context_engine.domain.model.Chunk;
+import com.kairos.context_engine.domain.model.content.Chunk;
 
 import java.util.List;
 import java.util.UUID;
