+++ src/main/java/com/kairos/context_engine/infrastructure/relational/repository/source/SpringSourceRepositoryAdapter.java
@@ -1,6 +1,6 @@
 package com.kairos.context_engine.infrastructure.relational.repository.source;
 
-import com.kairos.context_engine.domain.model.Source;
+import com.kairos.context_engine.domain.model.content.Source;
 import com.kairos.context_engine.domain.port.repository.SourceRepository;
 import com.kairos.context_engine.infrastructure.relational.entity.SourceEntity;
 import org.springframework.data.domain.PageRequest;
