+++ src/test/java/com/kairos/context_engine/infrastructure/semantic/SemanticSearchAdapterTest.java
@@ -1,6 +1,6 @@
 package com.kairos.context_engine.infrastructure.semantic;
 
-import com.kairos.context_engine.domain.model.Chunk;
+import com.kairos.context_engine.domain.model.content.Chunk;
 import com.kairos.context_engine.infrastructure.relational.entity.ChunkEntity;
 import com.kairos.context_engine.infrastructure.relational.entity.SourceEntity;
 import com.kairos.context_engine.infrastructure.relational.repository.chunk.JpaChunkRepository;
