+++ src/test/java/com/kairos/context_engine/use_case/GenerateSourceContextUseCaseTest.java
@@ -5,9 +5,9 @@ import com.kairos.context_engine.application.use_case.GenerateSourceContextUseCa
 import com.kairos.context_engine.domain.port.embedding.EmbeddingProvider;
 import com.kairos.context_engine.domain.port.graph.KnowledgeGraphStore;
 import com.kairos.context_engine.domain.port.extraction.TripleExtractor;
-import com.kairos.context_engine.domain.model.Chunk;
+import com.kairos.context_engine.domain.model.content.Chunk;
 import com.kairos.context_engine.domain.model.KnowledgeTriple;
-import com.kairos.context_engine.domain.model.Source;
+import com.kairos.context_engine.domain.model.content.Source;
 import com.kairos.context_engine.domain.model.Triple;
 import com.kairos.context_engine.domain.port.repository.ChunkRepository;
 import com.kairos.context_engine.domain.port.repository.SourceRepository;
