+++ src/test/java/com/kairos/context_engine/use_case/SearchSourceUseCaseTest.java
@@ -2,6 +2,8 @@ package com.kairos.context_engine.use_case;
 
 import com.kairos.context_engine.application.query.SearchSourceQuery;
 import com.kairos.context_engine.application.use_case.SearchSourceUseCase;
+import com.kairos.context_engine.domain.model.content.Chunk;
+import com.kairos.context_engine.domain.model.content.Source;
 import com.kairos.context_engine.domain.port.embedding.EmbeddingProvider;
 import com.kairos.context_engine.domain.port.graph.KnowledgeGraphSearch;
 import com.kairos.context_engine.domain.model.*;
